refactor(audio): extract SplitCsvLine to CsvParser to remove concrete CliRunner coupling
