+++ src/HaPcRemote.Core/Services/CliRunner.cs
@@ -1,5 +1,4 @@
 using System.Diagnostics;
-using System.Text;
 
 namespace HaPcRemote.Service.Services;
 
@@ -55,42 +54,5 @@ public class CliRunner : ICliRunner
         }
 
         return stdout;
-    }
-
-    public static List<string> SplitCsvLine(string line)
-    {
-        var fields = new List<string>();
-        var current = new StringBuilder();
-        var inQuotes = false;
-
-        for (var i = 0; i < line.Length; i++)
-        {
-            var c = line[i];
-
-            if (c == '"')
-            {
-                if (inQuotes && i + 1 < line.Length && line[i + 1] == '"')
-                {
-                    current.Append('"');
-                    i++;
-                }
-                else
-                {
-                    inQuotes = !inQuotes;
-                }
-            }
-            else if (c == ',' && !inQuotes)
-            {
-                fields.Add(current.ToString());
-                current.Clear();
-            }
-            else
-            {
-                current.Append(c);
-            }
-        }
-
-        fields.Add(current.ToString());
-        return fields;
     }
 }
